chore: move index to deprecated folder
